requirements.txt
Django>=1.7
pymysql==0.10.1
SQLAlchemy==1.4.19
protobuf==3.3.0
django-rest-swagger==2.2.0
gevent~=21.1.2
utils~=0.9.0
djangorestframework~=3.9.4